Rename components/LanguageSelector.vue to components/_shared/SideNavigation/LanguageSelector.vue
