+++ components/_shared/SideNavigation.vue
@@ -156,8 +156,8 @@ import {
 import Vue from 'vue'
 import { TranslateResult } from 'vue-i18n'
 
+import LanguageSelector from '@/components/_shared/SideNavigation/LanguageSelector.vue'
 import AppLink from '@/components/AppLink.vue'
-import LanguageSelector from '@/components/LanguageSelector.vue'
 import MenuList from '@/components/MenuList.vue'
 
 type Item = {
